refactor(node): move build_node to node.py
