+++ awaitwhat/dot.py
@@ -1,34 +1,12 @@
 import asyncio
-import io
 import json
 import re
 
 from dataclasses import dataclass
 
 from .blocker import blockers
-from .stack import task_print_stack
-from .node import Node
+from .node import Node, build_node
 from .utils import concise_stack_trace, concise_other
-
-
-def build_node(task, current):
-    if isinstance(task, asyncio.Task):
-        buf = io.StringIO()
-        data = task_print_stack(task, None, buf)
-        try:
-            name = task.get_name()
-        except AttributeError:
-            name = "Task"
-        # FIXME can we ever see a "done" task?
-        if task.done():
-            state = "done"
-        else:
-            state = "current" if task is current else "pending"
-        return Node(name, state, buf.getvalue(), task)
-    elif isinstance(task, asyncio.Future):
-        return Node("Future", None, None, task)
-    else:
-        return Node(str(task), None, None, task)
 
 
 def describe(task, current=None):
