+++ awaitwhat/node.py
@@ -1,4 +1,8 @@
+import asyncio
+import io
 from dataclasses import dataclass
+from .stack import task_print_stack
+
 
 @dataclass
 class Node:
@@ -8,3 +12,21 @@ class Node:
     task: str = None
 
 
+def build_node(task, current):
+    if isinstance(task, asyncio.Task):
+        buf = io.StringIO()
+        data = task_print_stack(task, None, buf)
+        try:
+            name = task.get_name()
+        except AttributeError:
+            name = "Task"
+        # FIXME can we ever see a "done" task?
+        if task.done():
+            state = "done"
+        else:
+            state = "current" if task is current else "pending"
+        return Node(name, state, buf.getvalue(), task)
+    elif isinstance(task, asyncio.Future):
+        return Node("Future", None, None, task)
+    else:
+        return Node(str(task), None, None, task)
